python/src/requirements.txt
Flask==2.3.2
gunicorn==20.1.0
google-cloud-bigquery==3.11.1
google-cloud-storage==2.9.0